+++ src/_stub.rs
